Now using smart pointers
Changed implementation of colors to use shared_ptr, renamed some files, added a new example.
